frontend/flyout: fix path of opening file in log by re-using handle_log_click and re-use border style
